improved responsiveness of branch section
